Importance of Smart Cities
Acting ethically is more than upholding core values or looking out for your stakeholders while conducting your work. Ethical data science and technology development demands rigor, honesty, excellent communication skills, the ability to embrace uncertainty, and a willingness to acknowledge mistakes or failures.

What my project was first about...

My main goal was to focus on trying to organize data on the homeless in our community. I want to bring attention that the homeless is kind of a condition that someone goes through and not an intended habit, some of these people just have no way of getting out of it and I feel as if they need help to be pushed out of their conditions because it all starts with technology. 
I think if we're able to centralize where the traffic is coming from we can control areas where there are homeless people and be able to locate them more efficiently through data collecting process and eventually help my partners locate homeless people and have them get the support and care they need to be back on their feet.

The Weaknesses I saw in my project goal...

Instead of trying to locate areas where homeless people are coming from which I wasn’t able to do because of the ethical aspects of data and its confidentiality to those people. 
Plus my project partners were restricted from giving me that data for obvious reasons considering these people don’t want their specific information being thrown at random people to look at.
I realized that Non-profit organizations collect data mainly for the purpose of being funded by the government

What My Project was Finally about...

My main goal was to still focus on trying to organize data on the homeless in our community and bring attention that the homeless is kind of a condition that someone goes through and not an intended habit. What I wanted to do was to Create a system of data where we can recognize whether homeless people are ending up in permanent Destinations, Temporary Destinations, Institutional settings, or Other locations. 
Being able to give P.E.H. the insights they aren’t getting from the data they are collecting as they never touch it after getting the funding they need from the government. 

I had the opportunity to work with Partners Ending Homelessness(PEH), which is a primary planning and coordinating body for homeless housing and services in Monroe County, NY.